Refactored metronome and volume components
