+++ src/routes/App.css
@@ -3,37 +3,6 @@
   font-family: 'Montserrat';
 }
 
-#logoHolder {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  top: 30%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-#splashText {
-  color: #575757;
-  font-size: 50px;
-  font-weight: bold;
-  position: absolute;
-  bottom: 5px;
-}
-
-#splashLogo {
-  width: 400px;
-}
-
-#splashBottom {
-  display: flex;
-  flex: 1;
-  margin: 50px;
-  justify-content: space-between;
-  align-items: flex-end;
-}
-
 #left {
   font-size: 50px;
   font-weight: bold;
@@ -49,27 +18,6 @@
 #right .material {
   margin-right: 20px;
   font-size: 50px;
-}
-
-#splash {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  display: flex;
-  background: rgba(177, 177, 177, 0.34);
-  backdrop-filter: blur(80px);
-  z-index: 2000;
-  color: white;
-  align-items: flex-end;
-  transition: opacity 0.9s;
-}
-
-.removed {
-  opacity: 0;
-}
-
-.none {
-  display: none !important;
 }
 
 .App {
@@ -89,10 +37,6 @@
   align-items: center;
   width: 100%;
   justify-content: space-between;
-}
-
-#metronome {
-  color: white;
 }
 
 #container {
@@ -161,15 +105,6 @@
 .keyPressed {
   animation-name: keyPress;
   animation-duration: 0.4s;
-}
-
-input[type=text] {
-  background: none;
-  color: white;
-  font-family: inherit;
-  font-size: 20px;
-  width: fit-content;
-  border: none;
 }
 
 @keyframes keyPress {
@@ -298,10 +233,6 @@ input[type=text] {
 
 }
 
-#bpmInput {
-  margin-left: 20px;
-}
-
 #note {
   align-self: flex-start;
   width: 300px;
@@ -340,83 +271,4 @@ input[type=text] {
        -moz-user-select: none;
         -ms-user-select: none;
             user-select: none;
-}
-
-#metronome {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-#metroIcon {
-  width: 25px;
-}
-
-#beats {
-  font-size: 20px;
-  margin-left: 3px;
-  margin-right: 20px;
-}
-
-#play {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 31px;
-  height: 31px;
-  left: 298px;
-  top: 247px;
-
-  background: rgba(87, 129, 192, 0.77);
-  box-shadow: 0px 0px 14px #5781C0;
-  border-radius: 5px;
-}
-
-#volumeIcon {
-  margin-right: 10px;
-  font-size: 30px;
-}
-
-#volumeWrap {
-  display: flex;
-  align-items: center;
-}
-
-.slider {
-  z-index: 100;
-  -webkit-appearance: none;
-  width: 150px;
-  height: 20px;
-  border-radius: 5px;
-  outline: none;
-  opacity: 0.7;
-  -webkit-transition: .2s;
-  transition: opacity .2s;
-
-  background: #6D7689;
-  border-radius: 16px;
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25), inset 0px 0px 3px 1px rgba(0, 0, 0, 0.53);
-}
-
-.slider:hover {
-  opacity: 1;
-}
-
-.slider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 25px;
-  height: 25px;
-  border-radius: 50%;
-  background: white;
-  cursor: pointer;
-}
-
-.slider::-moz-range-thumb {
-  width: 25px;
-  height: 25px;
-  border-radius: 50%;
-  background: white;
-  cursor: pointer;
 }
